+++ examples/pick_list/src/main.rs
@@ -43,10 +43,10 @@ impl Sandbox for Example {
         .placeholder("Choose a language...");
 
         let content = column![
-            vertical_space(600),
+            vertical_space().height(600),
             "Which is your favorite language?",
             pick_list,
-            vertical_space(600),
+            vertical_space().height(600),
         ]
         .width(Length::Fill)
         .align_items(Alignment::Center)